Move some test mods into appropriate packages (#5006)
