+++ src/test/java/net/minecraftforge/debug/gameplay/GiveItemToPlayerTest.java
@@ -17,7 +17,7 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  */
 
-package net.minecraftforge.debug;
+package net.minecraftforge.debug.gameplay;
 
 import net.minecraft.item.ItemStack;
 import net.minecraft.util.EnumActionResult;
