+++ src/test/java/net/minecraftforge/debug/world/WorldProviderCanSleepAtTest.java
@@ -17,7 +17,7 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  */
 
-package net.minecraftforge.debug;
+package net.minecraftforge.debug.world;
 
 import org.apache.logging.log4j.Logger;
 import net.minecraft.entity.player.EntityPlayer;
@@ -29,8 +29,8 @@ import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.common.Mod.EventHandler;
 import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
 
-@Mod(modid = CanSleepAtTest.MODID, name = "CanSleepAtTest", version = "0.0.0", acceptableRemoteVersions = "*")
-public class CanSleepAtTest
+@Mod(modid = WorldProviderCanSleepAtTest.MODID, name = "Can Sleep At Test", version = "0.0.0", acceptableRemoteVersions = "*")
+public class WorldProviderCanSleepAtTest
 {
     public static final String MODID = "can_sleep_at_test";
     public static final boolean ENABLED = false;
@@ -56,7 +56,7 @@ public class CanSleepAtTest
         @Override
         public DimensionType getDimensionType()
         {
-            return CanSleepAtTest.dimType;
+            return WorldProviderCanSleepAtTest.dimType;
         }
 
         @Override
